[CMYK-118] 대화 삭제 기능

+++ src/main/java/com/cmyk/ego/speaktoyouspring/api/hub/user_account/UserAccountDTO.java
@@ -14,7 +14,7 @@ public class UserAccountDTO {
     @NotBlank(message = "UID는 필수입니다.")
     private String uid;
 
-    private Integer egoId;
+    private Long egoId;
 
     @NotBlank(message = "email은 필수입니다.")
     private String email;
